--- a/test/rules/no-focused-tests.js
+++ b/test/rules/no-focused-tests.js
@@ -17,37 +17,93 @@ eslintTester.run('no-focused-tests', rule, {
       code: 'ddescribe("My exclusive suite", function() {});',
       errors: [
         {
-          message: 'Unexpected ddescribe.',
+          message: 'Unexpected ddescribe("My exclusive suite")',
           type: 'CallExpression'
         }
       ]
     },
+    // TODO: enable this when parsing template literals succeeds
+    // Saw:
+    //
+    // AssertionError: A fatal parsing error occurred: Parsing error:
+    // Unexpected character '`'
+//    {
+//      code: 'ddescribe(`My exclusive ${ suiteName }`, function() {});',
+//      errors: [
+//        {
+//          message: 'Unexpected ddescribe(`My exclusive ${ suiteName }`)',
+//          type: 'CallExpression'
+//        }
+//      ]
+//    },
     {
       code: 'iit("My exclusive test", function() {});',
       errors: [
         {
-          message: 'Unexpected iit.',
+          message: 'Unexpected iit("My exclusive test")',
           type: 'CallExpression'
         }
       ]
     },
+    // TODO: enable this when parsing template literals succeeds
+    // Saw:
+    //
+    // AssertionError: A fatal parsing error occurred: Parsing error:
+    // Unexpected character '`'
+//    {
+//      code: 'iit(`My exclusive ${ testName }`, function() {});',
+//      errors: [
+//        {
+//          message: 'Unexpected iit(`My exclusive ${ testName }`)',
+//          type: 'CallExpression'
+//        }
+//      ]
+//    },
     {
       code: 'fdescribe("My focused suite", function() {});',
       errors: [
         {
-          message: 'Unexpected fdescribe.',
+          message: 'Unexpected fdescribe("My focused suite")',
           type: 'CallExpression'
         }
       ]
     },
+    // TODO: enable this when parsing template literals succeeds
+    // Saw:
+    //
+    // AssertionError: A fatal parsing error occurred: Parsing error:
+    // Unexpected character '`'
+//    {
+//      code: 'fdescribe(`My exclusive ${ suiteName }`, function() {});',
+//      errors: [
+//        {
+//          message: 'Unexpected fdescribe(`My exclusive ${ suiteName }`)',
+//          type: 'CallExpression'
+//        }
+//      ]
+//    },
     {
       code: 'fit("My focused spec", function() {});',
       errors: [
         {
-          message: 'Unexpected fit.',
+          message: 'Unexpected fit("My focused spec")',
           type: 'CallExpression'
         }
       ]
     }
+    // TODO: enable this when parsing template literals succeeds
+    // Saw:
+    //
+    // AssertionError: A fatal parsing error occurred: Parsing error:
+    // Unexpected character '`'
+//    {
+//      code: 'fit(`My exclusive ${ testName }`, function() {});',
+//      errors: [
+//        {
+//          message: 'Unexpected fit(`My exclusive ${ testName }`)',
+//          type: 'CallExpression'
+//        }
+//      ]
+//    },
   ]
 })