@tlvince I managed to add test/suite name reporting for 2 rules using the prohibit helper.

Could you please take a look at why can't it parse template literals?
I tested it in one of my projects with the same prohibit code as I posted here and it works.