@burabure 2016-11-05
Hey @jrencz could you provide some context on the feature or fix this merge addresses?