@jrencz 2016-11-06
I think this feature is needed because when there's a lot of tests in the project (by a lot I mean thousands or unit tests or maybe hundreds if they are long running tests for example in case one uses Jasmine in selenium test suites) it becomes a common practice to either fdescribe the part you're currently working on or xdescribe tests you know the part you're working on can't affect.

Later (before pushing changes) it's needed to undo those temporary changes and in case the description is included in eslint error messages it's a bit easier to find them.

Real life use case: I have no-focused-tests and no-disabled-tests rules enabled with level error in the CI environment, but in dev environment they are both warnings (because of reasons described above - it's convenient)

It sometimes happens that either me or my teammates push without re-enabling tests. With clear and detailed description of what has been forced or disabled in CI build failure error is easier for other person to fix this silly mistake instead of waiting for the author to fix it in case everything else is merge-ready