@burabure 2016-11-07
@jrencz ok I understand now, this would be useful indeed. have you [benchmarked the performance hit](http://eslint.org/docs/developer-guide/working-with-rules#per-rule-performance) between the original implementation and this one?