@jrencz 2016-11-09
`➜  eslint-plugin-jasmine git:(master) ✗ TIMING=1 ./node_modules/.bin/eslint --no-eslintrc --rule "no-focused-tests: [2]" --rulesdir lib/rules lib`

Rule             | Time (ms) | Relative
:----------------|----------:|--------:
no-focused-tests |     0.902 |   100.0%
no-focused-tests |     0.816 |   100.0%
no-focused-tests |     0.682 |   100.0%
no-focused-tests |     0.732 |   100.0%
no-focused-tests |     0.720 |   100.0%
no-focused-tests |     1.225 |   100.0%
no-focused-tests |     0.772 |   100.0%
no-focused-tests |     0.846 |   100.0%
no-focused-tests |     0.698 |   100.0%
no-focused-tests |     0.676 |   100.0%

Average: 0.8069
Stddev: 0.1648

`➜  eslint-plugin-jasmine git:(master) ✗ TIMING=1 ./node_modules/.bin/eslint --no-eslintrc --rule "no-disabled-tests: [2]" --rulesdir lib/rules lib`

Rule              | Time (ms) | Relative
:-----------------|----------:|--------:
no-disabled-tests |     0.761 |   100.0%
no-disabled-tests |     0.732 |   100.0%
no-disabled-tests |     0.704 |   100.0%
no-disabled-tests |     0.683 |   100.0%
no-disabled-tests |     0.665 |   100.0%
no-disabled-tests |     0.708 |   100.0%
no-disabled-tests |     0.692 |   100.0%
no-disabled-tests |     0.668 |   100.0%
no-disabled-tests |     0.763 |   100.0%
no-disabled-tests |     0.782 |   100.0%

Average: 0.7158
Stddev: 0.04166


`➜  eslint-plugin-jasmine git:(prohibit-better-reporting) ✗ TIMING=1 ./node_modules/.bin/eslint --no-eslintrc --rule "no-focused-tests: [2]" --rulesdir lib/rules lib`

Rule             | Time (ms) | Relative
:----------------|----------:|--------:
no-focused-tests |     0.804 |   100.0%
no-focused-tests |     0.822 |   100.0%
no-focused-tests |     0.844 |   100.0%
no-focused-tests |     0.819 |   100.0%
no-focused-tests |     0.914 |   100.0%
no-focused-tests |     0.871 |   100.0%
no-focused-tests |     0.880 |   100.0%
no-focused-tests |     1.307 |   100.0%
no-focused-tests |     0.919 |   100.0%
no-focused-tests |     0.812 |   100.0%

Average: 0.8992
Stddev: 0.14914

`➜  eslint-plugin-jasmine git:(prohibit-better-reporting) ✗ TIMING=1 ./node_modules/.bin/eslint --no-eslintrc --rule "no-disabled-tests: [2]" --rulesdir lib/rules lib`

Rule              | Time (ms) | Relative
:-----------------|----------:|--------:
no-disabled-tests |     0.879 |   100.0%
no-disabled-tests |     0.872 |   100.0%
no-disabled-tests |     0.796 |   100.0%
no-disabled-tests |     0.838 |   100.0%
no-disabled-tests |     0.800 |   100.0%
no-disabled-tests |     0.801 |   100.0%
no-disabled-tests |     0.832 |   100.0%
no-disabled-tests |     0.866 |   100.0%
no-disabled-tests |     0.890 |   100.0%
no-disabled-tests |     0.876 |   100.0%

Average: 0.845
Stddev: 0.03635


Rule                    | Time on master (ms) | Time on prohibit-better-reporting (ms) | Impact
:----------------|------:|----:|--------:
no-focused-tests  | 0.8069 | 0.8992  | +11,4%
no-disabled-tests | 0.7158  | 0.845     | + 18%