@burabure 2016-11-16
seems reasonable enough, what do you think @lukeapage ?

@jrencz is this ready for final review on your side?