@jrencz 2016-11-17
@burabure i wasn't able to make all tests work (as mentioned in the description). I need some help with this